requirements.txt
click==7.1.2
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1
injector==0.18.3
numpy==1.19.1
passlib==1.7.2
pydantic==1.6.1
PyJWT==1.7.1
python-multipart==0.0.5
six==1.15.0
Sphinx==2.4.3
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
